pastry: completely overhaul data recording
This commit makes the recording system far more flexible, adds some
additional fields to recordings, and instead records individual worlds
instead of the entire runtime of the game.

The recording format now consists of a header followed by a list of
events. More event types can now be added in the future. As for the
changes to the recorded data:

  - The recording header contains a version field, string lookup table,
    and the number of events stored in the recording (no need to read
    until EOF.)
  - String data (entity/blockentity names and auxiliary data) is now
    stored in a lookup table at the start of the recording.
  - The frame event no longer includes entities and block entities,
    which have been given their own events.
  - The frame event now records the position and rotation of the player
    camera.
  - The entity event now records the position of the entity.
  - The entity event now stores the "path" (entity ID) instead of the
    full translation key.
  - The block entity event now records the position of the block entity.
  - The block entity event now stores the data portion in a separate
    field instead of including it in the name field.

The necessary changes to pastry-cli appear in the next commit. I plan on
moving the analysis features to the Pastry mod and leaving pastry-cli
primarily as a debugging/correctness checking tool.