change: capture time instead of frame number
I don't consider this a breaking change since the binary format of
PastryCaptureFrameEvent didn't change at all.

Storing the frame number is useless now that every (unpaused) frame is
captured. The time elapsed since the start of the capture is much more
useful, since it provides information about frame pacing which can be
used to e.g. provide a visual replay of the user's actions.