feat: add basic preemptive analysis
Currently, the preemptive analysis package is only capable of extracting
"consistent" preemptive readings (those with the same options, entities,
and blockentities throughout) and splitting them into valid and invalid
readings (those which have a silverfish spawner and those which don't.)

I would like to improve the consistency filter so that it is less
restrictive (e.g. a few frames with different characteristics are fine
if they are discarded or do not heavily shift the results.) I would also
like to add more actual analysis features.